Extract OrgFilterButton component
