+++ localization/react-intl/src/app/components/feed/FeedTopBar.json
@@ -1,27 +1,7 @@
 [
   {
-    "id": "feedTopBar.showItems",
-    "description": "Tooltip message displayed on button that the user presses in order to show items from an organization.",
-    "defaultMessage": "Show items from {orgName}"
-  },
-  {
-    "id": "feedTopBar.hideItems",
-    "description": "Tooltip message displayed on button that the user presses in order to hide items from an organization.",
-    "defaultMessage": "Hide items from {orgName}"
-  },
-  {
-    "id": "feedTopBar.noListSelected",
-    "description": "Message displayed on feed top bar when there is no list associated with the feed.",
-    "defaultMessage": "no list selected"
-  },
-  {
-    "id": "feedTopBar.customList",
-    "description": "Tooltip message displayed on button that the user presses in order to navigate to the custom list page.",
-    "defaultMessage": "Go to custom list"
-  },
-  {
     "id": "feedTopBar.addOrg",
-    "description": "Tooltip message displayed on a button that takes the user toa page where they can add an organization to this shared feed with an expectation to collaborate with the organization.",
+    "description": "Tooltip message displayed on a button that takes the user to a page where they can add an organization to this shared feed with an expectation to collaborate with the organization.",
     "defaultMessage": "Add a collaborating organization"
   }
 ]
