+++ localization/react-intl/src/app/components/feed/OrgFilterButton.json
@@ -1,0 +1,22 @@
+[
+  {
+    "id": "feedTopBar.showItems",
+    "description": "Tooltip message displayed on button that the user presses in order to show items from an organization.",
+    "defaultMessage": "Show items from {orgName}"
+  },
+  {
+    "id": "feedTopBar.hideItems",
+    "description": "Tooltip message displayed on button that the user presses in order to hide items from an organization.",
+    "defaultMessage": "Hide items from {orgName}"
+  },
+  {
+    "id": "feedTopBar.noListSelected",
+    "description": "Message displayed on feed top bar when there is no list associated with the feed.",
+    "defaultMessage": "no list selected"
+  },
+  {
+    "id": "feedTopBar.customList",
+    "description": "Tooltip message displayed on button that the user presses in order to navigate to the custom list page.",
+    "defaultMessage": "Go to custom list"
+  }
+]
